+++ tests/e2e/e2e.go
@@ -114,19 +114,6 @@ func getKubeConfig() *rest.Config {
 	return config
 }
 
-func installVelero() {
-	kubeConfig := getKubeConfig()
-
-	// create dynamic client for CR
-	client, err := dynamic.NewForConfig(kubeConfig)
-	if err != nil {
-		panic(err)
-	}
-	// get Velero unstruct type to create Velero CR
-	unstrVel := decodeYaml()
-	createVeleroCR(unstrVel, client)
-}
-
 func isPodRunning() wait.ConditionFunc {
 	return func() (bool, error) {
 		kubeConf := getKubeConfig()
