+++ tests/e2e/install.go
@@ -1,0 +1,18 @@
+package e2e
+
+import (
+	"k8s.io/client-go/dynamic"
+)
+
+func installVelero() {
+	kubeConfig := getKubeConfig()
+
+	// create dynamic client for CR
+	client, err := dynamic.NewForConfig(kubeConfig)
+	if err != nil {
+		panic(err)
+	}
+	// get Velero unstruct type to create Velero CR
+	unstrVel := decodeYaml()
+	createVeleroCR(unstrVel, client)
+}
