+++ tests/e2e/install_test.go
@@ -1,20 +1,9 @@
 package e2e
 
 import (
-	"testing"
-
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
-
-func TestVeleroE2E(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Velero E2E Suite")
-}
-
-var _ = BeforeSuite(func() {
-	installVelero()
-})
 
 var _ = Describe("Installing Velero", func() {
 	Context("When the default Velero CR is created", func() {
